backend/pyproject.toml
[project]
name = "e-exam-backend"
version = "0.1.0"
description = "E-exam-prepare FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    # Database
    "sqlalchemy>=2.0.37",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.14.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0",
    "email-validator>=2.2.0",
    # HTTP client (for RAG service)
    "httpx>=0.28.0",
    # Async task queue
    "celery>=5.4.0",
    "redis>=5.2.0",
    # Utilities
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
